feat(editor): handle visual meta highlighting
