+++ desktop/Cargo.toml
@@ -16,5 +16,6 @@ syntect = "5"
 once_cell = "1"
 chrono = { version = "0.4", default-features = false, features = ["clock"] }
 pulldown-cmark = "0.9"
+regex = "1"
 
 # app modules: state, actions, view
