+++ desktop/src/editor/mod.rs
@@ -1,8 +1,9 @@
 pub mod code_editor;
 pub mod autocomplete;
+pub mod syntax_highlighter;
 
 pub use code_editor::CodeEditor;
-pub use code_editor::THEME_SET;
+pub use syntax_highlighter::THEME_SET;
 pub use autocomplete::{AutocompleteState, Suggestion, suggestions};
 
 #[cfg(test)]
